+++ src/customEditors/EditorSensorItem.tsx
@@ -4,6 +4,7 @@ import Sensor from '../types/Sensor';
 
 import produce from 'immer';
 import { ColorDot } from 'components/ColorDot';
+import { MappingsInput } from 'components/MappingsInput';
 
 interface Props {
   sensor: Sensor;
@@ -31,9 +32,10 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
       {/* <HorizontalGroup> */}
       <Field label="Query ID" description="Set this as query ID OR the query alias below">
         <Input
+          css=""
           value={sensor.query.id}
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.query.id = event.currentTarget.value;
             });
           }}
@@ -42,9 +44,10 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
 
       <Field label="Query Alias" description="If both alias and ID are set, ID has precedence">
         <Input
+          css=""
           value={sensor.query.alias}
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.query.alias = event.currentTarget.value;
             });
           }}
@@ -52,12 +55,16 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
       </Field>
       {/* </HorizontalGroup> */}
 
-      <Field label="Mapping ID" description="Past here the ID of the mapping you want to use for this sensor">
-        <Input
-          value={sensor.mappingId}
-          onChange={event => {
-            updateSensor(sensor => {
-              sensor.mappingId = event.currentTarget.value;
+      <Field
+        label="Mapping IDs"
+        description="Select IDs of mappings you want to use for this sensor. First valid mapping will be applied. List can be reordered by dragging."
+      >
+        <MappingsInput
+          css=""
+          mappings={sensor.mappingIds}
+          onChange={(mappings) => {
+            updateSensor((sensor) => {
+              sensor.mappingIds = mappings;
             });
           }}
         />
@@ -65,10 +72,23 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
 
       <Field label="Name">
         <Input
+          css=""
           value={sensor.name}
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.name = event.currentTarget.value;
+            });
+          }}
+        />
+      </Field>
+
+      <Field label="Icon" description="Write a valid FontAwesome icon">
+        <Input
+          css=""
+          value={sensor.iconName}
+          onChange={(event) => {
+            updateSensor((sensor) => {
+              sensor.iconName = event.currentTarget.value;
             });
           }}
         />
@@ -76,9 +96,10 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
 
       <Field label="Link">
         <Input
+          css=""
           value={sensor.link}
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.link = event.currentTarget.value;
             });
           }}
@@ -87,9 +108,10 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
 
       <Field label="Show">
         <Switch
+          css=""
           value={sensor.visible}
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.visible = event.currentTarget.checked;
             });
           }}
@@ -99,8 +121,8 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
       <Field label="Unit Type">
         <UnitPicker
           value={sensor.unit}
-          onChange={unit => {
-            updateSensor(sensor => {
+          onChange={(unit) => {
+            updateSensor((sensor) => {
               sensor.unit = unit;
             });
           }}
@@ -111,12 +133,11 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
       <Field label="Font Color">
         <ColorPicker
           color={sensor.fontColor}
-          onChange={color => {
-            updateSensor(sensor => {
+          onChange={(color) => {
+            updateSensor((sensor) => {
               sensor.fontColor = color;
             });
           }}
-          enableNamedColors
         >
           {({ ref, showColorPicker, hideColorPicker }) => (
             <Button ref={ref} onMouseLeave={hideColorPicker} onClick={showColorPicker} variant="secondary">
@@ -129,12 +150,11 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
       <Field label="Background Color">
         <ColorPicker
           color={sensor.backgroundColor}
-          onChange={color => {
-            updateSensor(sensor => {
+          onChange={(color) => {
+            updateSensor((sensor) => {
               sensor.backgroundColor = color;
             });
           }}
-          enableNamedColors
         >
           {({ ref, showColorPicker, hideColorPicker }) => (
             <Button ref={ref} onMouseLeave={hideColorPicker} onClick={showColorPicker} variant="secondary">
@@ -147,10 +167,11 @@ export const EditorSensorItem: React.FC<Props> = (props: Props) => {
 
       <Field label="Decimals">
         <Input
+          css=""
           value={sensor.decimals}
           type="number"
-          onChange={event => {
-            updateSensor(sensor => {
+          onChange={(event) => {
+            updateSensor((sensor) => {
               sensor.decimals = Number.parseInt(event.currentTarget.value, 10);
             });
           }}