+++ CHANGELOG.md
@@ -1,9 +1,24 @@
+# v1.0.6
+
+## Fix
+
+* Remove enableNamedColors, which caused some colors to not render correctly - thanks [peishaofeng](https://github.com/peishaofeng)
+* UI Slowdown on v8.3.3 - thanks [rvdwijngaard](https://github.com/rvdwijngaard)
+
+# v1.0.5
+
+## Improvements
+
+* Multiple mappings support
+* Override value on mapping
+
 # v1.0.4
 
 ## Improvements
 
 * Variable support on sensor name
 * Variable support on sensor link
+* Add support for Grafana v8+
 
 # v1.0.3